cambio de librerias

+++ holberton.h
@@ -1,14 +1,20 @@
 #ifndef mpf
 #define mpf
 #include <stdarg.h>
+#include <string.h>
+#include <stdlib.h>
+#include <stdio.h>
+#include <stdarg.h>
+#include <unistd.h>
+#include <limits.h>
 int _printf(const char *format, ...);
 int mngfmt(const char *c, va_list);
 int prt_char(va_list);
 int prt_str(va_list);
 int prt_int(va_list);
 int prt_mod(va_list);
-char *itoch(int number);
-int cntdtsn(int number);
+char *itoch(long number);
+int cntdtsn(long number);
 /**
  * struct prt_format - struct of printers operators
  * @t: Is the character.
